move all code generation and target boilerplate to core
